pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "computer-use-engine"
version = "0.1.0"
description = "AI Computer Use agent framework — MCP server for desktop automation"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "yonghwan1106" }]
keywords = ["mcp", "computer-use", "automation", "ai-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Desktop Environment",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp>=1.25.0",
    "pyautogui>=0.9.54",
    "Pillow>=11.0.0",
    "pywin32>=311",
    "pygetwindow>=0.0.9",
    "pyperclip>=1.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
cue = "cue.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["cue"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
